+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2MapperExtensions.java
@@ -24,15 +24,18 @@ import com.concordium.grpc.v2.Policy;
 import com.concordium.grpc.v2.ProtocolVersion;
 import com.concordium.grpc.v2.ReleaseSchedule;
 import com.concordium.grpc.v2.*;
+import com.concordium.sdk.crypto.bls.BLSPublicKey;
 import com.concordium.sdk.crypto.bulletproof.BulletproofGenerators;
 import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
 import com.concordium.sdk.crypto.elgamal.ElgamalPublicKey;
 import com.concordium.sdk.crypto.pedersencommitment.PedersenCommitmentKey;
 import com.concordium.sdk.crypto.pointchevalsanders.PSPublicKey;
+import com.concordium.sdk.crypto.vrf.VRFPublicKey;
 import com.concordium.sdk.requests.AccountQuery;
 import com.concordium.sdk.requests.BlockQuery;
 import com.concordium.sdk.requests.EpochQuery;
 import com.concordium.sdk.responses.Epoch;
+import com.concordium.sdk.responses.FinalizerIndex;
 import com.concordium.sdk.responses.Round;
 import com.concordium.sdk.responses.TimeoutParameters;
 import com.concordium.sdk.responses.*;
@@ -42,6 +45,9 @@ import com.concordium.sdk.responses.accountinfo.*;
 import com.concordium.sdk.responses.accountinfo.credential.CredentialType;
 import com.concordium.sdk.responses.accountinfo.credential.*;
 import com.concordium.sdk.responses.bakersrewardperiod.BakerInfo;
+import com.concordium.sdk.responses.blockcertificates.QuorumSignature;
+import com.concordium.sdk.responses.blockcertificates.SuccessorProof;
+import com.concordium.sdk.responses.blockcertificates.TimeoutSignature;
 import com.concordium.sdk.responses.blockitemsummary.AccountCreationDetails;
 import com.concordium.sdk.responses.blockitemsummary.*;
 import com.concordium.sdk.responses.blocksummary.FinalizationData;
@@ -58,6 +64,8 @@ import com.concordium.sdk.responses.chainparameters.ChainParametersV1;
 import com.concordium.sdk.responses.chainparameters.ChainParametersV2;
 import com.concordium.sdk.responses.chainparameters.CooldownParametersCpv1;
 import com.concordium.sdk.responses.chainparameters.*;
+import com.concordium.sdk.responses.consensusstatus.BlockSignature;
+import com.concordium.sdk.responses.consensusstatus.ConsensusDetailedStatus;
 import com.concordium.sdk.responses.consensusstatus.ConsensusStatus;
 import com.concordium.sdk.responses.election.ElectionInfoBaker;
 import com.concordium.sdk.responses.poolstatus.BakerPoolStatus;
@@ -101,6 +109,180 @@ import static com.google.common.collect.ImmutableList.copyOf;
  * Object Mapping Extensions. Maps from GRPC types to client types and vice versa.
  */
 interface ClientV2MapperExtensions {
+
+    static com.concordium.grpc.v2.ConsensusDetailedStatusQuery to(final com.concordium.sdk.requests.ConsensusDetailedStatusQuery input) {
+        return com.concordium.grpc.v2.ConsensusDetailedStatusQuery.newBuilder()
+                .setGenesisIndex(GenesisIndex.newBuilder().setValue(input.getGenesisIndex()))
+                .build();
+    }
+
+    static ConsensusDetailedStatus to(final com.concordium.grpc.v2.ConsensusDetailedStatus input) {
+        return ConsensusDetailedStatus.builder()
+                .genesisBlock(Hash.from(input.getGenesisBlock()))
+                .persistentRoundStatus(to(input.getPersistentRoundStatus()))
+                .roundStatus(to(input.getRoundStatus()))
+                .nonFinalizedTransactionCount(UInt64.from(input.getNonFinalizedTransactionCount()))
+                .transactionTablePurgeCounter(input.getTransactionTablePurgeCounter())
+                .blockTable(to(input.getBlockTable()))
+                .branches(to(input.getBranchesList(), ClientV2MapperExtensions::to))
+                .roundExistingBlocks(to(input.getRoundExistingBlocksList(), ClientV2MapperExtensions::to))
+                .roundExistingQCs(to(input.getRoundExistingQcsList(), ClientV2MapperExtensions::to))
+                .genesisBlockHeight(com.concordium.sdk.types.AbsoluteBlockHeight.from(input.getGenesisBlockHeight().getValue()))
+                .lastFinalizedBlock(Hash.from(input.getLastFinalizedBlock()))
+                .lastFinalizedBlockHeight(input.getLastFinalizedBlockHeight().getValue())
+                .epochBakers(to(input.getEpochBakers()))
+                .timeoutMessages(to(input.getTimeoutMessages()))
+                .terminalBlock(Hash.from(input.getTerminalBlock()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RoundExistingQC to(final RoundExistingQC input) {
+        return com.concordium.sdk.responses.consensusstatus.RoundExistingQC.builder()
+                .round(Round.from(input.getRound()))
+                .epoch(Epoch.from(input.getEpoch()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RoundExistingBlock to(final RoundExistingBlock input) {
+        return com.concordium.sdk.responses.consensusstatus.RoundExistingBlock.builder()
+                .round(Round.from(input.getRound()))
+                .baker(to(input.getBaker()))
+                .block(Hash.from(input.getBlock()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.BranchBlocks to(final BranchBlocks input) {
+        return com.concordium.sdk.responses.consensusstatus.BranchBlocks.builder()
+                .blocksAtBranchHeight(to(input.getBlocksAtBranchHeightList(), Hash::from))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.PersistentRoundStatus to(final com.concordium.grpc.v2.PersistentRoundStatus input) {
+        return com.concordium.sdk.responses.consensusstatus.PersistentRoundStatus.builder()
+                .lastSignedQuorumMessage(to(input.getLastSignedQuorumMessage()))
+                .lastSignedTimeoutMessage(to(input.getLastSignedTimeoutMessage()))
+                .lastBakedRound(Round.from(input.getLastBakedRound()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.QuorumMessage to(final com.concordium.grpc.v2.QuorumMessage input) {
+        return com.concordium.sdk.responses.consensusstatus.QuorumMessage.builder()
+                .signature(QuorumSignature.from(input.getSignature()))
+                .block(Hash.from(input.getBlock()))
+                .finalizer(FinalizerIndex.from(input.getFinalizer()))
+                .round(Round.from(input.getRound()))
+                .epoch(Epoch.from(input.getEpoch()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.TimeoutMessage to(final com.concordium.grpc.v2.TimeoutMessage input) {
+        return com.concordium.sdk.responses.consensusstatus.TimeoutMessage.builder()
+                .finalizer(FinalizerIndex.from(input.getFinalizer()))
+                .round(Round.from(input.getRound()))
+                .epoch(Epoch.from(input.getEpoch()))
+                .quorumCertificate(to(input.getQuorumCertificate()))
+                .signature(TimeoutSignature.from(input.getSignature()))
+                .messageSignature(BlockSignature.from(input.getMessageSignature()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RawQuorumCertificate to(final com.concordium.grpc.v2.RawQuorumCertificate input) {
+        return com.concordium.sdk.responses.consensusstatus.RawQuorumCertificate.builder()
+                .blockHash(Hash.from(input.getBlockHash()))
+                .round(Round.from(input.getRound()))
+                .epoch(Epoch.from(input.getEpoch()))
+                .aggregatedSignature(QuorumSignature.from(input.getAggregateSignature()))
+                .signatories(to(input.getSignatoriesList(), FinalizerIndex::from))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RawFinalizationEntry to(final com.concordium.grpc.v2.RawFinalizationEntry input) {
+        return com.concordium.sdk.responses.consensusstatus.RawFinalizationEntry.builder()
+                .finalizedQc(to(input.getFinalizedQc()))
+                .successorQc(to(input.getSuccessorQc()))
+                .successorProof(SuccessorProof.from(input.getSuccessorProof()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.BakersAndFinalizers to(final com.concordium.grpc.v2.BakersAndFinalizers input) {
+        return com.concordium.sdk.responses.consensusstatus.BakersAndFinalizers.builder()
+                .bakers(to(input.getBakersList(), ClientV2MapperExtensions::to))
+                .finalizers(to(input.getFinalizersList(), com.concordium.sdk.responses.BakerId::from))
+                .bakerTotalStake(CCDAmount.fromMicro(input.getBakerTotalStake().getValue()))
+                .finalizerTotalStake(CCDAmount.fromMicro(input.getFinalizerTotalStake().getValue()))
+                .finalizationCommitteeHash(Hash.from(input.getFinalizationCommitteeHash().getValue().toByteArray()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.FullBakerInfo to(final com.concordium.grpc.v2.FullBakerInfo input) {
+        return com.concordium.sdk.responses.consensusstatus.FullBakerInfo.builder()
+                .bakerIdentity(com.concordium.sdk.responses.BakerId.from(input.getBakerIdentity()))
+                .electionVerifyKey(VRFPublicKey.from(input.getElectionVerifyKey().getValue().toByteArray()))
+                .signatureVerifyKey(ED25519PublicKey.from(input.getSignatureVerifyKey().getValue().toByteArray()))
+                .aggregationVerifyKey(BLSPublicKey.from(input.getAggregationVerifyKey().getValue().toByteArray()))
+                .stake(CCDAmount.fromMicro(input.getStake().getValue()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.EpochBakers to(com.concordium.grpc.v2.EpochBakers input) {
+        return com.concordium.sdk.responses.consensusstatus.EpochBakers.builder()
+                .previousEpochBakers(to(input.getPreviousEpochBakers()))
+                .currentEpochBakers(to(input.getCurrentEpochBakers()))
+                .nextEpochBakers(to(input.getNextEpochBakers()))
+                .nextPayday(to(input.getNextPayday()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.TimeoutMessages to(com.concordium.grpc.v2.TimeoutMessages input) {
+        return com.concordium.sdk.responses.consensusstatus.TimeoutMessages.builder()
+                .firstEpoch(Epoch.from(input.getFirstEpoch()))
+                .firstEpochTimeouts(to(input.getFirstEpochTimeoutsList(), ClientV2MapperExtensions::to))
+                .secondEpochTimeouts(to(input.getSecondEpochTimeoutsList(), ClientV2MapperExtensions::to))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.BlockTableSummary to(com.concordium.grpc.v2.BlockTableSummary input) {
+        return com.concordium.sdk.responses.consensusstatus.BlockTableSummary.builder()
+                .deadBlockCacheSize(UInt64.from(input.getDeadBlockCacheSize()))
+                .liveBlocks(to(input.getLiveBlocksList(), Hash::from))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RawFinalizerRound to(RawFinalizerRound input) {
+        return com.concordium.sdk.responses.consensusstatus.RawFinalizerRound.builder()
+                .round(Round.from(input.getRound()))
+                .finalizers(to(input.getFinalizersList(), FinalizerIndex::from))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RawTimeoutCertificate to(RawTimeoutCertificate input) {
+        return com.concordium.sdk.responses.consensusstatus.RawTimeoutCertificate.builder()
+                .round(Round.from(input.getRound().getValue()))
+                .minEpoch(Epoch.from(input.getMinEpoch()))
+                .qcRoundsFirstEpoch(to(input.getQcRoundsFirstEpochList(), ClientV2MapperExtensions::to))
+                .qcRoundsSecondEpoch(to(input.getQcRoundsSecondEpochList(), ClientV2MapperExtensions::to))
+                .aggregateSignature(TimeoutSignature.from(input.getAggregateSignature()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RoundTimeout to(RoundTimeout input) {
+        return com.concordium.sdk.responses.consensusstatus.RoundTimeout.builder()
+                .timeoutCertificate(to(input.getTimeoutCertificate()))
+                .quorumCertificate(to(input.getQuorumCertificate()))
+                .build();
+    }
+
+    static com.concordium.sdk.responses.consensusstatus.RoundStatus to(RoundStatus input) {
+        return com.concordium.sdk.responses.consensusstatus.RoundStatus.builder()
+                .currentRound(Round.from(input.getCurrentRound()))
+                .highestCertifiedBlock(to(input.getHighestCertifiedBlock()))
+                .previousRoundTimeout(to(input.getPreviousRoundTimeout()))
+                .roundEligibleToBake(input.getRoundEligibleToBake())
+                .currentEpoch(Epoch.from(input.getCurrentEpoch()))
+                .lastEpochFinalizationEntry(to(input.getLastEpochFinalizationEntry()))
+                .currentTimeout(to(input.getCurrentTimeout()))
+                .build();
+    }
 
     static com.concordium.grpc.v2.BlockHashInput to(final BlockQuery input) {
         BlockHashInput.Builder builder = BlockHashInput.newBuilder();